cli/requirements.txt
beautifulsoup4
faiss-cpu
langchain==0.1.11
langchain-google-genai
langchain-google-vertexai
langchain-openai
pypdf
pytest
pytest-mock
python-dotenv
tiktoken
typer[all]
typing_extensions
typing-inspect